% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mly_tavg_normal.R
\docType{data}
\name{mly_tavg_normal}
\alias{mly_tavg_normal}
\title{U.S. Monthly Climate Normals (1981-2010)}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4740 rows and 3 columns.}
\source{
\url{https://doi.org/10.7289/V5PN93JP}
}
\usage{
mly_tavg_normal
}
\description{
The U.S. Monthly Climate Normals for 1981 to 2010 are 30-year averages of meteorological parameters for thousands of U.S. stations located across the 50 states, as well as U.S. territories, commonwealths, the Compact of Free Association nations, and one station in Canada. NOAA Climate Normals are a large suite of data products that provide users with many tools to understand typical climate conditions for thousands of locations across the United States. As many NWS stations as possible are used, including those from the NWS Cooperative Observer Program (COOP) Network as well as some additional stations that have a Weather Bureau Army-Navy (WBAN) station identification number, including stations from the Climate Reference Network (CRN). The comprehensive U.S. Climate Normals dataset includes various derived products including daily air temperature normals (including maximum and minimum temperature normal, heating and cooling degree day normal, and others), precipitation normals (including snowfall and snow depth, percentiles, frequencies and other), and hourly normals (all normal derived from hourly data including temperature, dew point, heat index, wind chill, wind, cloudiness, heating and cooling degree hours, pressure normals). In addition to the standard set of normals, users also can find "agricultural normals", which are used in many industries, including but not limited to construction, architecture, pest control, etc. These supplemental "agricultural normals" include frost-freeze date probabilities, growing degree day normals, probabilities of reaching minimum temperature thresholds, and growing season length normals. Users can access the data either by product or by station. Included in the dataset is extensive documentation to describe station metadata, filename descriptions, and methodology of producing the data. All data utilized in the computation of the 1981-2010 Climate Normals were taken from the ISD Lite (a subset of derived Integrated Surface Data), the Global Historical Climatology Network-Daily dataset, and standardized monthly temperature data (COOP). These source datasets (including intermediate datasets used in the computation of products) are also archived at the NOAA NCDC.
}
\keyword{datasets}
